dominios
classes de dominios iniciadas

+++ src/main/java/com/vinicius/helpdesk/domain/enums/Perfil.java
@@ -22,7 +22,7 @@ public enum Perfil {
         return descricao;
     }
 
-    public static Perfil toEnum(Long cod) {
+    public static Perfil toEnum(Integer cod) {
         if (cod == null) {
             return null;
         }
